requirements.txt
fastapi
uvicorn
pdfplumber
python-docx
spacy
nltk
scikit-learn
pandas
numpy
python-multipart
jinja2
python-dotenv
sentence-transformers
torch

